pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aviutl2-api"
version = "0.1.0"
description = "Python API for AviUtl ver.2 project files (.aup2)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "swatanabe" }
]
keywords = ["aviutl", "video-editing", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pillow>=10.0",
    "numpy>=1.24",
    "opencv-python>=4.8",
]

[project.scripts]
aviutl2 = "aviutl2_api.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aviutl2_api"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
